+++ MandelbrotViewer/AppVersion.cs
@@ -8,8 +8,8 @@ namespace MandelbrotViewer;
 public static class AppVersion
 {
     public const int Major = 2;
-    public const int Minor = 3;
-    public const int Patch = 15;
+    public const int Minor = 5;
+    public const int Patch = 14;
 
     public static string Full => $"{Major}.{Minor}.{Patch}";